Updated micros to run off text input.
This adds a new implementation of schemaParser which can match CSV files
to input (there is no bounds checking, but java should throw an
OutOfBoundsException for us). The rest of the checkin uses this new
schemaParser to allow us to use simple text files for input rather than
the synthetic sources which capped us at 8kB of input. This also
restricts output at the sink to one line, which helps speed up the test
in the case of very large input.